event-management/pyproject.toml
[project]
name = "event-management"
version = "0.1.0"
description = "Event ingest from Pipecat; optional fan-out to client APIs."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "pyjwt>=2.10.0",
    "cryptography>=46.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
event-management = "app.main:run"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
